move palette dtos under folder
