--- dotnet-backend/Core/Dtos/AssignTagsToPaletteReq.cs
@@ -1,5 +1,0 @@
-public class AssignTagsToPaletteReq
-{
-    public List<int> ImageIds { get; set; } = new List<int>();
-    public int ProjectId { get; set; }
-}
