+++ dotnet-backend/Core/Dtos/PaletteService/AssignAssetsToProjectsReq.cs
@@ -1,6 +1,6 @@
-namespace Core.Dtos
+namespace Core.Dtos.PaletteService
 {
-    public class ModifyTagsReq
+    public class AssignAssetsToProjectsReq
     {
         public required string name { get; set; }
         public required string type { get; set; }
