+++ dotnet-backend/Core/Dtos/PaletteService/AssignTagsToAssetsReq.cs
@@ -1,6 +1,6 @@
-namespace Core.Dtos
+namespace Core.Dtos.PaletteService
 {
-    public class AssignAssetsToProjectsReq
+    public class AssignTagsToAssetsReq
     {
         public required string name { get; set; }
         public required string type { get; set; }
