+++ dotnet-backend/Core/Dtos/PaletteService/AssignTagsToPaletteReq.cs
@@ -1,0 +1,5 @@
+public class AssignTagsToPaletteReq
+{
+    public List<int> ImageIds { get; set; } = new List<int>();
+    public int ProjectId { get; set; }
+}
