+++ dotnet-backend/Core/Dtos/PaletteService/ModifyTagsReq.cs
@@ -1,6 +1,6 @@
-namespace Core.Dtos
+namespace Core.Dtos.PaletteService
 {
-    public class AssignTagsToAssetsReq
+    public class ModifyTagsReq
     {
         public required string name { get; set; }
         public required string type { get; set; }
