+++ dotnet-backend/Core/Dtos/PaletteService/UploadAssetsReq.cs
@@ -1,4 +1,4 @@
-namespace Core.Dtos
+namespace Core.Dtos.PaletteService
 {
     public class UploadAssetsReq
     {
